The managed runtime and the host pack were 4,207 lines in two files
`managed_runtime.rs` was 2,239: starting the guest, reaching the private
services inside it, warming the sandbox image, keeping the clock, and
deciding whether a missed probe is a fault. Eight modules --
`lifecycle`, `services`, `images`, `clock`, `probe`, `bootstrap`, `spec`
-- with the guards split the same way.

`native_host_pack.rs` was 1,968: where the app code is, this
installation's secrets, the stamps that decide whether a setup reruns,
and the manifest built from all of it. Five modules: `bindings`,
`secrets`, `stamps`, `build`.

Two more guards were reading files that had become directories, and one
of them mattered.
`a_release_build_ignores_every_runtime_redirecting_env_var` is the guard
that keeps a signed, notarized build from letting the environment choose
which runtime it fetches. It read locald's daemon directory and then
named five files beside it -- and four of those five are directories
now. It walks both crates from disk and asserts it found more than
thirty files.

The host pack's own two source guards had the same shape: one scanned
"the shipping half of the file", which meant everything above the
`#[cfg(test)]` in a file that no longer has one. Both read the pack's
modules from disk, minus the guards, so the scan cannot find its own
needles.
