+++ desktop/scripts/file-size-baseline.json
@@ -9,7 +9,5 @@
   "desktop/agent-host/tests/lemma_mcp_e2e.rs": 699,
   "desktop/agent-host/tests/real_harness_e2e.rs": 1202,
   "desktop/agent-host/tests/support/mod.rs": 1417,
-  "desktop/locald/src/agent_host.rs": 1564,
-  "desktop/locald/src/managed_runtime.rs": 2239,
-  "desktop/locald/src/native_host_pack.rs": 1968
+  "desktop/locald/src/agent_host.rs": 1564
 }
